Fix clippy lints
